requirements.txt
flask==3.0.0
docker==7.1.0
werkzeug==3.0.1
python-dotenv==1.0.0
flask-cors==4.0.0
urllib3==1.26.5
requests==2.31.0